Remove intrinsic binding directory
Removes the search for bindings in the intrinsic
directory defined at compile time.

Consequently, the option --no-ldpaths that was used
to remove the default intrinsic search of bindings
is removed.

It is still possible to automatically load bindings
through the environment variables AFB_LDPATHS and
AFB_WEAK_LDPATHS that if existing point to a colon
separated list of paths to recursively explore.