Add alias for /.well-known
By default, if the directory /usr/share/afb-binder/well-known
exists, it is served through HTTP as the /.well-known/ path of URLs.

The directory can be modified using the variables CMAKE_INSTALL_PREFIX,
DATADIR and WELL_KNOWN_DIR when invoking cmake.

If the environment variable AFB_BINDER_WELL_KNOWN_DIR is set, it is also
searched with precedence over the predefined directory.

Also note that existing .well-known directory in the root-http
served has precedence over the 2 others.